made renderers actually extendable
