+++ OceanSim/assets
@@ -1,1 +1,1 @@
-/home/julian/assets
+/home/julian/OceanSim_assets
